% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selUniProt.R
\name{checkValid}
\alias{checkValid}
\title{Visualize Uniprot feature keys}
\usage{
checkValid(infoPDB)
}
\arguments{
\item{accession}{Enter accession key.}

\item{pdbFile}{A PDB file, could be downloaded from PDB online}

\item{name}{the 4-letter PDB codes/identifiers of the PDB file
to be visualized.}
}
\value{
Displays the 3D model of the protein with the selected uniprot values.
}
\description{
Function that takes a PDB file and visualizes it as a 3D model. The user to
select on a scale to view uniprot feature keys
}
\examples{
# Using pdb file available with package
# ?idpr
#selUniProt("DNA", pdb, "1bm8")

}
\references{
Yao, X., G. Scarabelli, L. Skjaerven, and B. Grant (2020). Protein Structure Networks with Bio3D.
bio3D. http://thegrantlab.org/bio3d/articles/online/cna_vignette/cna_vignette.spin.html#references-1.

McFadden, W.M. (2020). Charge and Hydropathy Vignette.
https://rdrr.io/bioc/idpr/f/vignettes/chargeHydropathy-vignette.Rmd.
}
