% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdb}
\alias{pdb}
\title{PDB file of 1bm8}
\format{
List of 8 :
\describe{
\item{Atom}{List containing information on the atom and standard residues }
\item{XYZ}{atomic coordinate record containing the X,Y,Z}
\item{seqres}{List containing the Amino Acid sequence}
\item{Helix}{List indicating the location and type (right-handed alpha, etc.) of helices}
\item{Sheet}{List indicating the location, sense (anti-parallel, etc.)
and registration with respect to the previous strand in the sheet (if any) of each strand in the model}
\item{Remark}{identifies entries in which a re-refinement has been
performed using the data from an existing entry.}
\item{Call}{Referring to language/command used to get this}
}
}
\source{
PDB, https://www.rcsb.org/structure/1BM8
}
\usage{
pdb
}
\description{
Contains information on the protein 1bm8 gotten from PDB (Protein Data Bank).
}
\examples{
\dontrun{
 pdb
}
}
\keyword{datasets}
