% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdb2}
\alias{pdb2}
\title{PDB file of 1si4}
\format{
List of 7 (PDB file) :
\describe{
\item{Atom}{List containing information on the atom and standard residues }
\item{XYZ}{atomic coordinate record containing the X,Y,Z}
\item{seqres}{List containing the Amino Acid sequence}
\item{Helix}{List indicating the location and type (right-handed alpha, etc.) of helices}
\item{calpha}{Boolean of whether it contains the central carbon}
\item{Remark}{identifies entries in which a re-refinement has been
performed using the data from an existing entry.}
\item{Call}{Referring to language/command used to get this}
}
}
\source{
PDB, https://www.rcsb.org/structure/1si4
}
\usage{
pdb2
}
\description{
Contains information on the protein 1si4 gotten from PDB (Protein Data Bank).
}
\examples{
\dontrun{
 pdb2
}
}
\keyword{datasets}
