% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpdbSelectOcclude.R
\name{runpdbSelectOcclude}
\alias{runpdbSelectOcclude}
\title{Launch Shiny App for pdbSelectOcclude}
\usage{
runpdbSelectOcclude(pdbFile, name)
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for pdbSelectOcclude.
The purpose of this app is only to illustrate how a Shiny
app works. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

pdbSelectOcclude::runpdbSelectOcclude(pdb, "1bm8")
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
