% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selAA.R
\name{selAA}
\alias{selAA}
\title{Visualize occluded AA of PDB}
\usage{
selAA(index, pdbFile, name)
}
\arguments{
\item{index}{Valid residue index that refers to a AA code found in sequence.
The index can be found in showInfoPDB as the a number within the
range of code_unique.}

\item{pdbFile}{A PDB file, could be downloaded from PDB online}

\item{name}{the 4-letter PDB codes/identifiers of the PDB file
to be visualized.}
}
\value{
Returns chosen AA, and it displays the 3D model of the protein with
the selected AA visible.
}
\description{
Function that takes a PDB file and visualizes it as a 3D model. The user to select which AA to view.
This function allows user to interact and view occluded AA.
}
\examples{
# Using pdb file available with package
selAA(1, pdb, "1bm8")


}
\references{
Yao, X., G. Scarabelli, L. Skjaerven, and B. Grant (2020). Protein Structure Networks with Bio3D.
bio3D. http://thegrantlab.org/bio3d/articles/online/cna_vignette/cna_vignette.spin.html#references-1.

Su, W. (2021). Introduction to r3dmol. r3dmol.
https://cran.r-project.org/web/packages/r3dmol/vignettes/r3dmol.html.
}
