% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selHphob.R
\name{selHphobSlider}
\alias{selHphobSlider}
\title{Visualize polarity with Shiny}
\usage{
selHphobSlider(pdbFile, name)
}
\arguments{
\item{pdbFile}{A PDB file, could be downloaded from PDB online}

\item{name}{the 4-letter PDB codes/identifiers of the PDB file
to be visualized.}
}
\value{
None - it opens up shiny app with protein structure shown.
}
\description{
Function that takes a PDB file and visualizes it with Shiny.
Specifically for polarity.
}
\examples{
# Using pdb file available with package
#selHphobSlider(pdb, "1bm8")

}
\references{
Chang, W, et al. (2017). Using sliders. Shiny from Rstudio.
https://shiny.rstudio.com/articles/sliders.html.
}
