% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showInfoPDB.R
\name{showInfoPDB}
\alias{showInfoPDB}
\title{Show the information on chosen PDB file}
\usage{
showInfoPDB(pdbFile, name)
}
\arguments{
\item{pdbFile}{A PDB file, could be downloaded from PDB online}

\item{name}{the 4-letter PDB codes/identifiers of the PDB file
that was chosen as 'pdbFile' input}
}
\value{
Returns a list of containing 2 pieces information from the PDB file:
AA and chain information.
}
\description{
Function that takes a PDB file as input, and then extracts the AA (amino acid)
information, and chain information of the chosen protein.
}
\examples{
# Using pdb2 and pdb file available with package
showInfoPDB(pdb, "1bm8")

showInfoPDB(pdb2, "1si4")

}
\references{
Yao, X., G. Scarabelli, L. Skjaerven, and B. Grant (2020). Protein Structure Networks with Bio3D.
bio3D. http://thegrantlab.org/bio3d/articles/online/cna_vignette/cna_vignette.spin.html#references-1.

Yao, X., G. Scarabelli, L. Skjaerven, and B. Grant (2020). chain.pdb: Find Possible PDB Chain Breaks.
bio3D. https://rdrr.io/cran/bio3d/man/chain.pdb.html.
}
