% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visPDB.R
\name{visPDB}
\alias{visPDB}
\title{visualize a PDB file}
\usage{
visPDB(pdbfile)
}
\arguments{
\item{pdbfile}{A PDB file, could be downloaded from PDB online}
}
\value{
nothing for now. It just shows the 3D visuals
}
\description{
Function that visualizes the given PDB file
}
\references{
Yao, X., G. Scarabelli, L. Skjaerven, and B. Grant (2020). Protein Structure Networks with Bio3D. bio3D. http://thegrantlab.org/bio3d/articles/online/cna_vignette/cna_vignette.spin.html#references-1.
}
